Switch back to nvim-autopairs
mini.pairs handles pair-matching very well, except when text already
exists.

  if (foo == |this is a string")

type `"` and it becomes

  if (foo == ""this is a string")

mini.pairs checks a neighborhood of +/- 2 characters, and that doesn't
seem to be actually configurable. nvim-autopairs will, by default, scan
the entire line.

OTOH, the documentation for nvim-autopairs is atrocious. There is
literally no documentation for many of the options, no documentation of
the user API, and the rules DSL is.... silly.